Add files via upload
The Student Management System (SMS) is a robust, scalable, and efficient application developed using TypeScript. It is designed to streamline the administrative tasks of educational institutions by managing student data effectively. The system offers functionalities for adding students, enrolling in courses, paying fees, and status management, providing a comprehensive solution to handle the diverse needs of schools and universities.